refactor: useDragDrop step 1 — DOM position helpers
Creates src/hooks/useDragDrop.js and moves the 5 pure DOM position
helpers out of App.jsx: getHourHeight, minutesToPosition,
positionToMinutes, durationToHeight, calculateTaskPosition.

Hook is called after useTaskActions with calendarRef + timeGridRef.
All five helpers destructured back into App.jsx — no behaviour change.
